Added initial information about Pepade
Just basically threw in a place holder for it so people know it's there. I'm going to be playing with both the OS and the documentation a lot in the free time. Honestly, because it's so different from everything else we're running, it might be worth giving it it's own documentation.